fix: 예외 처리 추가
